src/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
amsterdam-schema-tools[django]==0.8.13  # via -r requirements.in
asgiref==3.2.10           # via django
attrs==19.3.0             # via jsonschema
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via amsterdam-schema-tools, mappyfile
defusedxml==0.6.0         # via django-gisserver
django-environ==0.4.5     # via amsterdam-schema-tools
django-gisserver==0.8.3   # via amsterdam-schema-tools
django-postgres-unlimited-varchar==1.1.0  # via amsterdam-schema-tools
django==3.0.8             # via -r requirements.in, amsterdam-schema-tools, django-gisserver, django-postgres-unlimited-varchar
geoalchemy2==0.8.4        # via amsterdam-schema-tools
idna==2.10                # via requests
jinja2==2.11.2            # via amsterdam-schema-tools
jsonref==0.2              # via mappyfile
jsonschema==3.2.0         # via amsterdam-schema-tools, mappyfile
lark-parser==0.7.8        # via mappyfile
mappyfile==0.8.4          # via amsterdam-schema-tools
markupsafe==1.1.1         # via jinja2
ndjson==0.3.1             # via amsterdam-schema-tools
orjson==3.2.0             # via django-gisserver
psycopg2==2.8.5           # via amsterdam-schema-tools
pyrsistent==0.16.0        # via jsonschema
python-dateutil==2.8.1    # via amsterdam-schema-tools
python-string-utils==1.0.0  # via -r requirements.in, amsterdam-schema-tools
pytz==2020.1              # via django
requests==2.24.0          # via amsterdam-schema-tools
shapely==1.7.0            # via amsterdam-schema-tools
six==1.15.0               # via jsonschema, python-dateutil
sqlalchemy==1.3.18        # via geoalchemy2
sqlparse==0.3.1           # via django
urllib3==1.25.9           # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
